backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.3
google-generativeai>=0.8.0
openai>=1.0.0
supabase>=2.13.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart>=0.0.6
